refactor: split routers and handler to separate files
