+++ src/pages/_app.tsx
@@ -1,10 +1,9 @@
 import { withTRPC } from '@trpc/next';
 import { AppType } from 'next/dist/shared/lib/utils';
-import { AppRouter } from './api/trpc/[trpc]';
+import { AppRouter } from '../utils/trpc/AppRouter';
 
-const MyApp: AppType = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />;
-};
+const MyApp: AppType = ({ Component, pageProps }) => 
+  <Component {...pageProps} />;
 
 export default withTRPC<AppRouter>({
   config({ ctx }) {
