+++ src/pages/api/trpc/[trpc].ts
@@ -1,22 +1,5 @@
-import * as trpc from '@trpc/server';
 import * as trpcNext from '@trpc/server/adapters/next';
-import { z } from 'zod';
-
-export const appRouter = trpc.router().query('hello', {
-  input: z
-    .object({
-      text: z.string().nullish(),
-    })
-    .nullish(),
-  resolve({ input }) {
-    return {
-      greeting: `hello ${input?.text ?? 'world'}`,
-    };
-  },
-});
-
-// export type definition of API
-export type AppRouter = typeof appRouter;
+import { appRouter } from '../../../utils/trpc/AppRouter';
 
 // export API handler
 export default trpcNext.createNextApiHandler({
