+++ src/utils/trpc/index.ts
@@ -1,5 +1,5 @@
 import { createReactQueryHooks } from '@trpc/react';
-import type { AppRouter } from '../pages/api/trpc/[trpc]';
+import type { AppRouter } from './AppRouter';
 
 export const trpc = createReactQueryHooks<AppRouter>();
 // => { useQuery: ..., useMutation: ...}
